feat: export page and table of contents components
