+++ package.json
@@ -44,7 +44,7 @@
     "@stoplight/json": "^3.0.3",
     "@stoplight/json-ref-resolver": "^2.1.1",
     "@stoplight/json-schema-viewer": "^2.5.1",
-    "@stoplight/markdown-viewer": "^3.3.1",
+    "@stoplight/markdown-viewer": "^3.4.0",
     "@stoplight/tree-list": "^4.8.0",
     "@stoplight/ui-kit": "^2.9.2",
     "@stoplight/yaml": "^3.0.1",
