+++ src/containers/Page.tsx
@@ -1,9 +1,9 @@
-import { Classes } from '@blueprintjs/core';
 import { safeParse } from '@stoplight/json';
 import cn from 'classnames';
 import * as React from 'react';
 
-import { Page as Component } from '../components/Page';
+import { Page as PageComponent } from '../components/Page';
+import { PageSkeleton } from '../components/PageSkeleton';
 import { useNodeInfo } from '../hooks/useNodeInfo';
 
 export interface IPage {
@@ -14,8 +14,6 @@ export interface IPage {
   className?: string;
   padding?: string;
 }
-
-export const PageComponent = Component;
 
 export const Page: React.FunctionComponent<IPage> = ({
   srn,
@@ -50,18 +48,5 @@ export const Page: React.FunctionComponent<IPage> = ({
       scrollInnerContainer={scrollInnerContainer}
       padding={padding}
     />
-  );
-};
-
-export const PageSkeleton: React.FunctionComponent<{ className?: string; padding?: string }> = ({
-  className,
-  padding = '10',
-}) => {
-  return (
-    <div className={cn('PageSkeleton', className, `p-${padding}`, 'flex flex-col h-full')}>
-      <div className={cn(Classes.SKELETON, 'h-12 w-1/5')} />
-      <div className={cn(Classes.SKELETON, 'h-12 my-6')} />
-      <div className={cn(Classes.SKELETON, 'flex-1 my-6')} />
-    </div>
   );
 };
