+++ src/index.ts
@@ -3,6 +3,11 @@ export * from './components/Docs';
 export * from './components/TryIt';
 export * from './components/HttpService';
 export * from './components/HttpOperation';
+export * from './components/PageSkeleton';
+export * from './components/TableOfContentsSkeleton';
+
+export { Page as PageComponent } from './components/Page';
+export { TableOfContents as TableOfContentsComponent } from './components/TableOfContents';
 
 export * from './containers/Hub';
 export * from './containers/Page';
